Fix: Link C Math Library, Add Library Build Option
**OVERVIEW**
This commit fixes an issue where the compiler flag `-lm` is required in
order for the build to work on Ubuntu Linux. In addition, it allows for
a boolean option `shared_lib` to be used to control whether or not a
shared library build is used.

**TEST PLAN**
This has been tested against a project using the meson build system.